+++ src/controllers/usuarios.js
@@ -1,41 +1,14 @@
 const express = require('express');
-const Usuario = require('../models/usuario_model');
-const Joi = require('@hapi/joi');
-const usuario_model = require('../models/usuario_model');
 const ruta = express.Router();
-
-//validaciones para el objeto usuario
-const schema = Joi.object({
-    nombre: Joi.string()
-        .min(3)
-        .max(30)
-        .required()
-        .pattern(/^[A-zA-záéíóú ]{3,30}$/),
-
-    password: Joi.string()
-        .pattern(/^[A-zA-Z0-9]{3,30}$/),
-
-    email: Joi.string()
-        .email({ minDomainSegments: 2, tlds: {allow: ['com', 'net', 'edu', 'co']}})
-});
-
-//función asíncrona para crear un objeto de tipo usuario
-async function crearUsuario(usuariox){
-    let usuario = new Usuario({
-        email       : usuariox.email,
-        nombre      : usuariox.nombre,
-        password    : usuariox.password
-    });
-return await usuario.save();
-}
+const logic = require('../logic/usuario_logic')
 
 // Endpoint de tipo POST para el recurso USUARIOS
 ruta.post('/',(req,res)=>{
     let body = req.body;
 
-    const {error, value} = schema.validate({nombre: body.nombre, email: body.email});
+    const {error, value} = logic.schema.validate({nombre: body.nombre, email: body.email});
     if(!error){
-        let resultado = crearUsuario(body);
+        let resultado = logic.crearUsuario(body);
 
         resultado.then( user => {
             res.json({
@@ -49,22 +22,11 @@ ruta.post('/',(req,res)=>{
     }
 });
 
-//función asíncrona para actualizar usuario
-async function actualizarUsuario(email, body){
-    let usuario = await Usuario.findOneAndUpdate({"email":email},{
-        $set: {
-            nombre : body.nombre,
-            password: body.password
-        }
-    }, {new: true});
-    return usuario;
-} 
-
 // End point de tipo PUT para actualizar los datos del usuario
 ruta.put('/:email', (req, res) => {
-    const {error, value} = schema.validate({nombre: req.body.nombre});
+    const {error, value} = logic.schema.validate({nombre: req.body.nombre});
     if(!error){
-        let resultado = actualizarUsuario(req.params.email, req.body);
+        let resultado = logic.actualizarUsuario(req.params.email, req.body);
         resultado.then(valor => {
             res.json({
                 valor
@@ -77,19 +39,9 @@ ruta.put('/:email', (req, res) => {
     }
 });
 
-//función asíncrona para inactivar un usuario
-async function desactivarUsuario(email){
-    let usuario = await Usuario.findOneAndUpdate({"email": email}, {
-        $set: {
-            estado: false
-        }
-    },{ new: true });
-    return usuario;
-}
-
 //End ponit de tipo DELETE para el recurso USUARIO
     ruta.delete('/:email', (req, res) => {
-        let resultado = desactivarUsuario(req.params.email);
+        let resultado = logic.desactivarUsuario(req.params.email);
         resultado.then(valor => {
             res.json({
                 usuario: valor
@@ -97,15 +49,9 @@ async function desactivarUsuario(email){
         });
     });
 
-//función asíncrona para listar todos los usuarios activos
-async function listarUsuariosActivos(){
-    let usuarios = await Usuario.find({"estado": true});
-    return usuarios;
-}
-
 //End point de tipo GET para el recurso usuarios. Lista todos los usuarios
 ruta.get('/',(req, res) =>{
-    let resultado = listarUsuariosActivos();
+    let resultado = logic.listarUsuariosActivos();
     resultado.then(usuarios =>{
         res.json(usuarios)
     }).catch(err => {
